# apps/sandbox-gateway/openapi.yaml
openapi: "3.0.3"
info:
  title: Lixeta Sandbox API
  version: "1.0.0"
  description: |
    The **Lixeta Sandbox API** is a real-time rules engine simulation platform.

    Submit domain events → receive decisions instantly, with full audit trails,
    revenue impact tracking, risk signals, and tamper-evident dispute packages.

    ## Authentication

    Protected endpoints accept an `X-Api-Key` header (`lxk_...` format).
    In development mode (`REQUIRE_AUTH=false`) authentication is disabled.

    ## Session lifecycle

    ```
    POST /session          → create a session, get sessionId
    POST /trigger-event    → submit events, get decisions
    GET  /analytics        → real-time KPIs
    GET  /logs             → full audit trail
    POST /dispute          → download evidence ZIP
    POST /session/close    → close session
    ```

    ## Response envelope

    All JSON responses use a consistent wrapper:
    ```json
    { "ok": true,  "requestId": "req_...", "timestamp": "...", "data": { ... } }
    { "ok": false, "requestId": "req_...", "timestamp": "...", "error": { "code": "...", "message": "..." } }
    ```
servers:
  - url: http://localhost:4000
    description: Local development server
security: []
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        API key in `lxk_<base64url>` format.
        Required when `REQUIRE_AUTH=true`. Disabled by default in development.
  schemas:
...
    ErrorResponse:
      type: object
      required: [ok, requestId, timestamp, error]
      properties:
        ok:
          type: boolean
          enum: [false]
          example: false
        requestId:
          type: string
          example: "req_abc123"
        timestamp:
          type: string
          format: date-time
          example: "2026-03-25T12:00:00.000Z"
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "sessionId is required."
            details:
              type: object
              additionalProperties: true
              description: Optional structured details (field errors, etc.)

    # ── Session ────────────────────────────────────────────────────────────────
    SessionMetadata:
      type: object
      description: Arbitrary metadata attached to a session.
      properties:
        label:
          type: string
          maxLength: 128
          example: "Sandbox Session"
        environment:
          type: string
          maxLength: 64
          example: "development"
        ownerId:
          type: string
          maxLength: 128
          example: "user_123"
        tags:
          type: array
          maxItems: 10
          items:
            type: string
            maxLength: 64
          example: ["sandbox-ui", "demo"]
    SessionCreatedData:
      type: object
      required: [sessionId, createdAt, expiresAt, status]
      properties:
        sessionId:
          type: string
          pattern: "^sess_[A-Za-z0-9_]+$"
          example: "sess_ABCDEFGHJKMNPabcde"
        createdAt:
          type: string
          format: date-time
          example: "2026-03-25T12:00:00.000Z"
        expiresAt:
          type: string
          format: date-time
          example: "2026-03-25T14:00:00.000Z"
        status:
          type: string
          enum: [active, completed, expired]
          example: "active"
        metadata:
          $ref: "#/components/schemas/SessionMetadata"
...
    RevenueEvent:
      type: object
      description: A revenue impact event produced by a rule action.
      properties:
        triggeringEventId:
          type: string
          example: "evt_req_abc123_s1"
        category:
          type: string
          example: "sms_saved"
        direction:
          type: string
          enum: [gain, loss]
          example: "gain"
        amount:
          type: object
          properties:
            amountMinorUnits:
              type: integer
              description: Amount in minor currency units (e.g. kobo for NGN).
              example: 5000
            currency:
              type: string
              example: "NGN"
        description:
          type: string
          example: "SMS notification suppressed — in-app delivery confirmed"
    RiskEvent:
      type: object
      description: A risk signal produced during engine evaluation.
      properties:
        triggeringEventId:
          type: string
          example: "evt_req_abc123_s1"
        category:
          type: string
          example: "geo_anomaly"
          description: One of geo_anomaly, velocity_breach, pattern_deviation, timezone_anomaly, sla_breach
        severity:
          type: string
          enum: [low, medium, high, critical]
          example: "high"
        score:
          type: number
          minimum: 0
          maximum: 1
          example: 0.82
        description:
          type: string
          example: "Device country (GB) does not match account country (NG)"
    RuleTraceCondition:
      type: object
      properties:
        field:
          type: string
          example: "payload.appOpen"
        operator:
          type: string
          example: "equals"
        expected:
          type: string
          example: "true"
        actual:
          type: string
          example: "true"
        passed:
          type: boolean
          example: true
    RuleTraceAction:
      type: object
      properties:
        actionType:
          type: string
          example: "route_to_in_app"
        executed:
          type: boolean
          example: true
    RuleTrace:
      type: object
      description: Per-rule execution trace from the engine.
      properties:
        ruleId:
          type: string
          example: "smart-notification"
        ruleName:
          type: string
          example: "Smart Notification Router"
        ruleVersion:
          type: string
          example: "1.0.0"
        triggeringEventId:
          type: string
          example: "evt_req_abc123_s1"
        outcome:
          type: string
          enum: [matched, not_matched, skipped, error]
          example: "matched"
        explanation:
          type: string
          example: "App is open — routing to in-app channel"
        executionTimeMs:
          type: integer
          example: 1
        conditions:
          type: array
          items:
            $ref: "#/components/schemas/RuleTraceCondition"
        actions:
          type: array
          items:
            $ref: "#/components/schemas/RuleTraceAction"
...
    SessionCounters:
      type: object
      properties:
        totalEventsProcessed:
          type: integer
          example: 12
        totalDecisions:
          type: integer
          example: 12
        allowCount:
          type: integer
          example: 9
        blockCount:
          type: integer
          example: 1
        flagCount:
          type: integer
          example: 2
        deferCount:
          type: integer
          example: 0
        errorCount:
          type: integer
          example: 0
    LogsData:
      type: object
      required: [session, pagination, events, revenueEvents, riskEvents, ruleTraces, decisions]
      properties:
        session:
          type: object
          properties:
            sessionId:
              type: string
              example: "sess_ABCDEFGHJKMNPabcde"
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
            expiresAt:
              type: string
              format: date-time
            status:
              type: string
              enum: [active, completed, expired]
              example: "active"
            counters:
              $ref: "#/components/schemas/SessionCounters"
        pagination:
          type: object
          properties:
            total:
              type: integer
              example: 12
            offset:
              type: integer
              example: 0
            limit:
              type: integer
              example: 100
            returned:
              type: integer
              example: 12
            hasMore:
              type: boolean
              example: false
        events:
          type: array
          description: Full DomainEvent objects in chronological order.
          items:
            type: object
            additionalProperties: true
        revenueEvents:
          type: array
          description: Revenue events correlated to the paginated events.
          items:
            $ref: "#/components/schemas/RevenueEvent"
        riskEvents:
          type: array
          description: Risk events correlated to the paginated events.
          items:
            $ref: "#/components/schemas/RiskEvent"
        ruleTraces:
          type: array
          description: Rule execution traces correlated to the paginated events.
          items:
            $ref: "#/components/schemas/RuleTrace"
        decisions:
          type: array
          description: Decision results correlated to the paginated events.
          items:
            type: object
            additionalProperties: true

    # ── Analytics ──────────────────────────────────────────────────────────────
...
paths:
...
  /session:
    post:
      operationId: createSession
      summary: Create simulation session
      description: |
        Opens a new session and returns a `sessionId`. All subsequent calls to
        `/trigger-event`, `/logs`, and `/analytics` require this ID.

        Sessions auto-expire after `ttlSeconds` (default 2 hours, maximum 24 hours).
        Use `POST /session/close` to close a session early.
      tags: [Session]
      security:
        - ApiKeyAuth: []
        - {}
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                metadata:
                  $ref: "#/components/schemas/SessionMetadata"
                ttlSeconds:
                  type: integer
                  minimum: 60
                  maximum: 86400
                  default: 7200
                  description: Session TTL in seconds (60–86400). Default 7200 (2 hours).
                  example: 7200
                scenarioId:
                  type: string
                  maxLength: 128
                  description: Optional scenario identifier for grouping test runs.
                  example: "diaspora_fraud_test"
                scenarioName:
                  type: string
                  maxLength: 256
                  description: Human-readable scenario name.
                  example: "Diaspora Fraud Detection — Velocity Test"
            example:
              metadata:
                label: "Sandbox Session"
                environment: "development"
                tags: ["sandbox-ui"]
              ttlSeconds: 7200
      responses:
        "201":
          description: Session created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    enum: [true]
                  requestId:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  data:
                    $ref: "#/components/schemas/SessionCreatedData"
              example:
                ok: true
                requestId: "req_abc123"
                timestamp: "2026-03-25T12:00:00.000Z"
                data:
                  sessionId: "sess_ABCDEFGHJKMNPabcde"
                  createdAt: "2026-03-25T12:00:00.000Z"
                  expiresAt: "2026-03-25T14:00:00.000Z"
                  status: "active"
                  metadata:
                    label: "Sandbox Session"
                    environment: "development"
                    tags: ["sandbox-ui"]
        "400":
          description: Invalid request body (bad metadata, invalid ttlSeconds, etc.)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          description: Session capacity exceeded. Retry after closing unused sessions.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /logs:
    get:
      operationId: getLogs
      summary: Retrieve session audit log
      description: |
        Returns all events, decisions, revenue events, risk events, and rule traces
        for a session — in chronological order.

        Supports pagination (`limit` / `offset`) and event type filtering.
        Accessing this endpoint increments the session's `accessCount` counter.
      tags: [Session]
      security:
        - ApiKeyAuth: []
        - {}
      parameters:
        - name: sessionId
          in: query
          required: true
          description: The session ID to retrieve logs for.
          schema:
            type: string
            pattern: "^sess_[A-Za-z0-9_]+$"
          example: "sess_ABCDEFGHJKMNPabcde"
        - name: limit
          in: query
          required: false
          description: Maximum number of events to return (1–1000, default 100).
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: offset
          in: query
          required: false
          description: Number of events to skip for pagination (default 0).
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: filter
          in: query
          required: false
          description: Filter by exact event type (e.g. `payment.initiated`).
          schema:
            type: string
          example: "payment.initiated"
      responses:
        "200":
          description: Session logs returned.
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    enum: [true]
                  requestId:
                    type: string
                  timestamp:
                    type: string
                    format: date-time
                  data:
                    $ref: "#/components/schemas/LogsData"
        "400":
          description: Missing or invalid query parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Session not found or expired.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Analytics ────────────────────────────────────────────────────────────────
...